Extract a new sql utility
